Verifying disks inertias
Disks with inertia equal to 0 would be a problem for the code interpretation. It could misslead the rotor type and guide the analyses through a completely different way it should.
Adding a inertia verification at the beginning will cut all disks with 0 inertia (mass only) for the rotor type consideration. Notice that it does not remove these disks from the model. They just won't be considered to the rotor type definition (if between bearings, overhung, etc).